Before moving function map to CoroutineUtil
